requirements.txt
guilded.py==1.13.1
aiohttp-client-cache==0.11.0
jikanpy-v4==1.0.2
pyyaml==6.0.1
motor==3.4.0
pydantic==2.7.1
beanie==1.26.0